refactor: move components to modules folder, improve code
