+++ src/modules/home/joinXplorers.tsx
@@ -1,4 +1,4 @@
-import { Button } from "../ui/button";
+import { Button } from "@/components/ui/button";
 
 const JoinXplorers = () => {
   return (
