+++ src/styles/globals.css
@@ -57,6 +57,6 @@ h6 {
     transform: translateX(0);
   }
   to {
-    transform: translateX(calc(-100% - 2.5rem));
+    transform: translateX(calc(-100% - 1.6rem));
   }
 }
